[feature] LottoUtils 생성 및 Generator 삭제
